Final commit
I finished the project fixed what was messed up in step 7 and commented the entire website. I do want to say that the site in the text book looks different from mine only because the max width is stoping it and the max width is what it is supposed to be.